refactor baseyear preprocessing
